+++ src/platform/windows/events_loop.rs
@@ -926,30 +926,31 @@ pub unsafe extern "system" fn callback(window: HWND, msg: UINT,
 
         winuser::WM_SETCURSOR => {
             let call_def_window_proc = CONTEXT_STASH.with(|context_stash| {
-                let cstash = context_stash.borrow();
-                let mut call_def_window_proc = false;
-                if let Some(cstash) = cstash.as_ref() {
-                    if let Some(w_stash) = cstash.windows.get(&window) {
-                        if let Ok(window_state) = w_stash.lock() {
-                            if window_state.mouse_in_window {
-                                match window_state.cursor_state {
-                                    CursorState::Normal => {
-                                        winuser::SetCursor(winuser::LoadCursorW(
-                                                ptr::null_mut(),
-                                                window_state.cursor));
-                                    },
-                                    CursorState::Grab | CursorState::Hide => {
-                                        winuser::SetCursor(ptr::null_mut());
-                                    }
+                context_stash
+                    .borrow()
+                    .as_ref()
+                    .and_then(|cstash| cstash.windows.get(&window))
+                    .map(|window_state_mutex| {
+                        let window_state = window_state_mutex.lock().unwrap();
+                        if window_state.mouse_in_window {
+                            match window_state.cursor_state {
+                                CursorState::Normal => {
+                                    let cursor = winuser::LoadCursorW(
+                                        ptr::null_mut(),
+                                        window_state.cursor.0,
+                                    );
+                                    winuser::SetCursor(cursor);
+                                },
+                                CursorState::Grab | CursorState::Hide => {
+                                    winuser::SetCursor(ptr::null_mut());
                                 }
-                            } else {
-                                call_def_window_proc = true;
                             }
+                            false
+                        } else {
+                            true
                         }
-                    }
-                }
-
-                call_def_window_proc
+                    })
+                    .unwrap_or(true)
             });
 
             if call_def_window_proc {